+++ NEWS.md
@@ -8,7 +8,7 @@ gridpattern v0.1
 
   * "regular_polygon" ``grid.pattern_regular_polygon()`` (#20)
 
-* Supports the following patterns from [ggpattern](https://github.com/coolbutuseless/ggpattern):
+* Supports the following patterns from [{ggpattern}](https://github.com/coolbutuseless/ggpattern):
 
   * "ambient" ``grid.pattern_ambient()``  
   * "circle" ``grid.pattern_circle()`` (#5)
@@ -23,13 +23,21 @@ gridpattern v0.1
   * [Custom ggpattern geometry-based patterns](https://coolbutuseless.github.io/package/ggpattern/articles/developing-patterns-2.html) (#1)
   * [Custom ggpattern array-based patterns](https://coolbutuseless.github.io/package/ggpattern/articles/developing-patterns-3.html) (#2)
 
-  Enhancements to the ``ggpattern`` patterns provided by this package:
+  Enhancements to the ``{ggpattern}`` patterns provided by this package:
 
     * Resolution of "array" patterns is now user-supplied.
       Defaults to 72 DPI but with a minimum of 12 pixel image width and/or height.
     * Default for "pattern_alpha" is now ``NA_real_`` (preserve existing alpha transparency) instead of ``1`` (set to fully opaque).
     * ``grid.pattern_ambient()`` pattern supports all arguments of the six [ambient](https://cran.r-project.org/package=ambient) 
       "noise" functions plus a "seed" argument to make them reproducible (for an identical image resolution).
+    * ``grid.pattern_circle()`` has several enhancements:
+
+      * Allows use of the argument ``xoffset`` and ``yoffset`` (#22)
+      * Allows use of argument ``type`` to setup a "hex" arrangement instead of default "square" arrangement.
+      * Use of multiple ``fill`` colors now produces a more attractive pattern.
+      * Use of ``density`` greater than ``1`` will now sometimes give an attractive result (#17).
+      * Uses ``{sf}`` to clip circles to boundary instead of ``{gridGeometry}``
+
     * ``grid.pattern_crosshatch()`` allows using the argument ``fill2`` to set 
        different fill color(s) for the "over" crosshatch lines (#14).